# optimize styleguide compilation time