--- a/lib/toolset/grunt-tasks/assembler/helpers.js
+++ b/lib/toolset/grunt-tasks/assembler/helpers.js
@@ -27,9 +27,22 @@ Helper = (function() {
    * @return {String} file's html compiled with jade
    */
   Helper.prototype.jadeRender = function (file, vars, extraTpl) {
-    var tpl;
+    var tpl = this.jadeCompile(file, extraTpl);
     vars = _.merge(vars, this.getJadeFunctions());
 
+    try {
+        return tpl(vars || {});
+    } catch (e) {
+        grunt.fail.fatal('Render error in ' + file + "\n" + e);
+    }
+  };
+
+  /**
+   * Utility function that receives a filepath and returns a Jade function.
+   * @param  {String} filepath
+   * @return {Function} a jade template function
+   */
+  Helper.prototype.jadeCompile = function (file, extraTpl) {
     try {
         tpl = require('jade').compile(grunt.file.read(file) + extraTpl, {
             filename    : file,
@@ -37,14 +50,9 @@ Helper = (function() {
             basedir     : srcDir,
             compileDebug: true
         });
+        return tpl;
     } catch (e) {
         grunt.fail.fatal('Build error in ' + file + "\n" + e);
-    }
-
-    try {
-        return tpl(vars || {});
-    } catch (e) {
-        grunt.fail.fatal('Render error in ' + file + "\n" + e);
     }
   };