--- a/lib/toolset/grunt-tasks/assembler/styleguideAssembler.js
+++ b/lib/toolset/grunt-tasks/assembler/styleguideAssembler.js
@@ -19,7 +19,9 @@ module.exports = function (grunt, options) {
       templatesPath = "lib/toolset/grunt-tasks/assembler/styleguide-templates/s-pages/components/",
       componentListTemplate         = templatesPath + "component-list.jade",
       componentTemplate             = templatesPath + "component.jade",
-      componentCombinationTemplate  = templatesPath + "component-combination.jade";
+      componentCombinationTemplate  = templatesPath + "component-combination.jade",
+      tplComponentCombination       = Helper.jadeCompile(componentCombinationTemplate),
+      tplComponent                  = Helper.jadeCompile(componentTemplate);
 
   /**
    * Given an array of arrays, generates all possible combinations
@@ -129,34 +131,38 @@ module.exports = function (grunt, options) {
   }
 
   function generateComponentPage(component, componentsData) {
+    var jadeFunctions = Helper.getJadeFunctions();
+    var vars = {};
+
     if (!component.hideInStyleguide) {
       component.combinations = generateComponentOptionsCombinations(component);
 
       grunt.log.write("Creating component " + component.name + " combinations.");
+
       _.each(component.combinations, function (combination, index) {
         grunt.log.write(".");
-        var combinationHtml = Helper.jadeRender(
-          componentCombinationTemplate,
-          {
-            "component": component,
-            "components": componentsData.allMetadata,
-            "combination": combination,
-            "relativePath": "../../../"
-          }
-        );
+
+        vars = _.merge({
+          "component": component,
+          "components": componentsData.allMetadata,
+          "combination": combination,
+          "relativePath": "../../../"
+        }, jadeFunctions);
+
+        var combinationHtml = tplComponentCombination(vars);
         grunt.file.write("dist/styleguide/components/" + component.name + "/c" + index + ".html", combinationHtml);
         // grunt.log.writeln("combination html"+ combinationHtml);
       });
 
       grunt.log.writeln("\nCreating component " + component.name + " page...");
-      var componentHtml = Helper.jadeRender(
-        componentTemplate,
-        {
-          "component": component,
-          "components": componentsData.allMetadata,
-          "relativePath": "../../"
-        }
-      );
+
+      vars = _.merge({
+        "component": component,
+        "components": componentsData.allMetadata,
+        "relativePath": "../../"
+      }, jadeFunctions);
+
+      var componentHtml = tplComponent(vars);
       grunt.file.write("dist/styleguide/components/" + component.name + ".html", componentHtml);
     }
   }