Tried to move the Jade template compilation out of the generateComponentPage function.

- Template are compiled to javascript function only once at the beginning.
- Now in `generateComponentPage` we simply call the generated template function instead of compiling it.

These are the results with the basler project when running `grunt stylguide`:

before:

    Execution Time (2015-04-28 07:39:55 UTC)
    styleguideAssembler  2m 44.5s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 100%
    Total 2m 45.2s

after:

    Execution Time (2015-04-28 07:48:04 UTC)
    loading tasks         1.3s  ▇▇▇▇▇▇▇▇▇▇▇ 26%
    clean:dist            89ms  ▇ 2%
    copy:images          107ms  ▇ 2%
    styleguideAssembler   3.5s  ▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇▇ 69%
    Total 5s